pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hfpef-registry-synth"
version = "0.1.0"
description = "Registry-first transparency-adjusted synthesis engine for HFpEF RCTs"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
  { name = "Codex" }
]
dependencies = [
  "requests>=2.31.0",
  "pandas>=2.0.0",
  "numpy>=1.24.0",
  "scipy>=1.10.0",
  "tqdm>=4.66.0",
  "python-dateutil>=2.8.2",
]

[project.optional-dependencies]
fuzzy = ["rapidfuzz>=3.0.0"]
graph = ["networkx>=3.0.0"]
test = ["pytest>=7.4.0"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
